+++ NEWS.md
@@ -11,6 +11,7 @@
   * `print.summary.sienaMeta`, `siena07`, `print01Report`: drop RForge revision.
   * Correct "objname" to "projname" in `meta.table` (`siena08.r`).
   * Simplify LaTeX output of `meta.table`. 
+  * `seq_along` and `seq_len` used in `print01Report`. 
 
 # RSiena 1.3.14