+++ NEWS.md
@@ -1,3 +1,18 @@
+# RSiena 1.3.15
+
+## 
+   
+2022-11-27
+
+## Changes in RSiena:  
+
+### Corrections:
+  * `siena08`: correct p-value `pTsq` for overall test statistic `Tsq`
+  * `print.summary.sienaMeta`, `siena07`, `print01Report`: drop RForge revision.
+  * Correct "objname" to "projname" in `meta.table` (`siena08.r`).
+  * Simplify LaTeX output of `meta.table`. 
+  * `seq_along` and `seq_len` used in `print01Report`. 
+
 # RSiena 1.3.14
 
 ##